backend/requirements-dev.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Database
SQLAlchemy==2.0.23
psycopg[binary]==3.1.12
alembic==1.13.1

# Validation
pydantic==2.6.4
pydantic-settings==2.2.1

# Background Tasks
celery==5.3.5
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Auth & Security
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Data Formats
PyYAML==6.0.1
orjson==3.9.13

# Observability
sentry-sdk==1.40.4
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# GCP / Vertex AI
google-cloud-aiplatform==1.47.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-pubsub==2.19.0

# RAG / Vector Store
chromadb==0.4.22

# Development & Testing
ruff==0.1.15
black==24.3.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-json-report==1.5.0

# Production server
gunicorn==21.2.0
